requirements.txt
zenoml>=0.4.0
inspiredco
sentence_transformers